+++ Changelog.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 2018-08-01
+### Added
+- Hashed each file in supplied directory and saved to its own results file
+
+### Changed
+- Updated output formatting 
+
 ## 2018-07-23
 ### Changed
 - Improved formatting for output file for vFeed results